The two UIs were formatting different numbers
Stage 4's acceptance criterion is a character-for-character match between
the Unity inspector and the Dash drawer, defended by a generated parity
fixture. It passes, and it has passed since session 18. It was checking the
wrong half: the formatters were correct on both sides, they were being
handed different numbers.

Python holds these values as binary64 and formats from that. The C# loader
parsed the same decimal text into binary32. Measured on a real 40-year,
3-deme export:

    stress  "-1.385"  ->  float64 -1.39   float32 -1.38   77 of 1323
    aerobic "39.305"  ->  float64  39.30  float32  39.31   64 of 1323

snapshots.py rounds stress and aerobic to three decimals and the drawer
prints two, so an exact 3-dp midpoint is the ORDINARY case and binary32
pushes about half of them the other way. epi_accel, viability, pedigree_f,
height and purity disagreed nowhere, and the reason is the same: their
stored precision is not finer than what is displayed. So about 5% of
villagers had a stress reading that differed between the two UIs in the
second decimal, both values entirely plausible, with nothing on either
screen to suggest a problem.

Fixed by parsing every column the viewer PRINTS as a double --
CsvParse.Double, new, alongside Float. Geometry (x, y) stays binary32
because it is never shown as text. Parsing the same decimal text to
binary64 on both sides is bit-identical, so the disagreement is gone
rather than made rarer. FrameRow grows 28 bytes, ~10 MB on a 600x600
world against a 300 MB budget.

Two guards, deliberately unlike each other. test_displayed_columns_are_
parsed_at_double_width is a source rule, because nothing that compares
formatters can see a reverted parse. And the C# test asserts BOTH
neighbours -- the right answer through Double and the wrong one through
Float -- so the bug and its fix sit next to each other and either can
fail. (The fixture is the third; it carries doubles as of the next
commit, which is where its new cases live.)

One break was caught by reading rather than compiling: ViabilityColor(0.9f)
widens to 0.89999997..., which IS less than 0.9, so an existing Stage-4
test would have flipped to CRIT. Every literal in those tests is now a
double.

The general lesson: a parity test pins the transformation, not the
pipeline. Ours compared two formatters and said nothing about the two
parsers feeding them.

+++ unity/com.samal.extnpc/Runtime/Data/WorldBundle.cs
@@ -150,17 +150,17 @@ namespace ExtNPC.Data
                         Age = CsvParse.Int(f[iAge]),
                         Deme = CsvParse.Int(f[iDeme]),
                         Lineage = pool.Get(f[iLin]),
-                        Purity = CsvParse.Float(f[iPur]),
+                        Purity = CsvParse.Double(f[iPur]),
                         Generation = CsvParse.Int(f[iGen]),
                         Children = CsvParse.Int(f[iKids]),
-                        Stress = CsvParse.Float(f[iStress]),
-                        EpiAccel = CsvParse.Float(f[iEpi]),
-                        Aerobic = CsvParse.Float(f[iAero]),
+                        Stress = CsvParse.Double(f[iStress]),
+                        EpiAccel = CsvParse.Double(f[iEpi]),
+                        Aerobic = CsvParse.Double(f[iAero]),
                         Conditions = CsvParse.Int(f[iCond]),
-                        PedigreeF = CsvParse.Float(f[iF]),
-                        Viability = CsvParse.Float(f[iViab]),
+                        PedigreeF = CsvParse.Double(f[iF]),
+                        Viability = CsvParse.Double(f[iViab]),
                         Cnv = CsvParse.Int(f[iCnv]),
-                        HeightCm = CsvParse.Float(f[iH]),
+                        HeightCm = CsvParse.Double(f[iH]),
                         LifeStage = pool.Get(f[iStage]),
                     };
                     if (!byTick.TryGetValue(tick, out var list))
@@ -365,8 +365,8 @@ namespace ExtNPC.Data
                         if (i == iTick || CsvParse.IsBlank(f[i])) continue;
                         // history.csv is all-numeric by construction; a column
                         // that is not is skipped rather than crashing the load.
-                        if (float.TryParse(f[i], NumberStyles.Float,
-                                CultureInfo.InvariantCulture, out float v))
+                        if (double.TryParse(f[i], NumberStyles.Float,
+                                CultureInfo.InvariantCulture, out double v))
                         {
                             row.Values[cols[i]] = v;
                         }
@@ -411,9 +411,9 @@ namespace ExtNPC.Data
                         Father = G("father"),
                         Deme = CsvParse.Int(G("deme")),
                         DemeLabel = G("deme_label"),
-                        PedigreeF = CsvParse.Float(G("pedigree_f")),
-                        RealisedF = CsvParse.Float(G("realised_f")),
-                        RelativeViability = CsvParse.Float(G("relative_viability")),
+                        PedigreeF = CsvParse.Double(G("pedigree_f")),
+                        RealisedF = CsvParse.Double(G("realised_f")),
+                        RelativeViability = CsvParse.Double(G("relative_viability")),
                         MendelianDiagnoses = CsvParse.List(G("mendelian_diagnoses")),
                         MendelianCarrierOf = CsvParse.List(G("mendelian_carrier_of")),
                         MedicalConditions = CsvParse.List(G("medical_conditions")),
